update the led control part
